+++ src/Exception/VideoException.php
@@ -25,22 +25,22 @@ class VideoException
         return self::exception(sprintf('The video height value is less than the permitted size: %d px', $height));
     }
 
-    static public function durationLarge(int $duration): Exception
+    static public function durationLarge(float $duration): Exception
     {
         return new Exception(Code::VIDEO_DURATION, sprintf('The video duration value is larger than the permitted: %d', $duration));
     }
 
-    static public function durationLess(int $duration): Exception
+    static public function durationLess(float $duration): Exception
     {
         return new Exception(Code::VIDEO_DURATION, sprintf('The video duration value is less than the permitted: %d', $duration));
     }
 
-    static public function bitrateLarge(int $bitrate): Exception
+    static public function bitrateLarge(float $bitrate): Exception
     {
         return new Exception(Code::VIDEO_BITRATE, sprintf('The video bitrate value is larger than the permitted: %d', $bitrate));
     }
 
-    static public function bitrateLess(int $bitrate): Exception
+    static public function bitrateLess(float $bitrate): Exception
     {
         return new Exception(Code::VIDEO_BITRATE, sprintf('The video bitrate value is less than the permitted: %d', $bitrate));
     }
